Enum cranelift_module::ModuleError

pub enum ModuleError {
    Undeclared(String),
    IncompatibleDeclaration(String),
    IncompatibleSignature(String, Signature, Signature),
    DuplicateDefinition(String),
    InvalidImportDefinition(String),
    Compilation(CodegenError),
    Backend(Error),
}

Error messages for all Module methods

Variants

Undeclared(String)

Indicates an identifier was used before it was declared

Tuple Fields of Undeclared

0: String

IncompatibleDeclaration(String)

Indicates an identifier was used as data/function first, but then used as the other

Tuple Fields of IncompatibleDeclaration

0: String

IncompatibleSignature(String, Signature, Signature)

Indicates a function identifier was declared with a different signature than declared previously

Tuple Fields of IncompatibleSignature

0: String1: Signature2: Signature

DuplicateDefinition(String)

Indicates an identifier was defined more than once

Tuple Fields of DuplicateDefinition

0: String

InvalidImportDefinition(String)

Indicates an identifier was defined, but was declared as an import

Tuple Fields of InvalidImportDefinition

0: String

Compilation(CodegenError)

Wraps a cranelift-codegen error

Tuple Fields of Compilation

0: CodegenError

Backend(Error)

Wraps a generic error from a backend

Tuple Fields of Backend

0: Error