Struct cranelift_module::ModuleDeclarations

pub struct ModuleDeclarations { /* fields omitted */ }

This provides a view to the state of a module which allows ir::ExternalNames to be translated into FunctionDeclarations and DataDeclarations.

Implementations

impl ModuleDeclarations

pub fn get_name(&self, name: &str) -> Option<FuncOrDataId>

Get the module identifier for a given name, if that name has been declared.

pub fn get_functions(
&self
) -> impl Iterator<Item = (FuncId, &FunctionDeclaration)>

Get an iterator of all function declarations

pub fn is_function(name: &ExternalName) -> bool

Return whether name names a function, rather than a data object.

pub fn get_function_decl(&self, func_id: FuncId) -> &FunctionDeclaration

Get the FunctionDeclaration for the function named by name.

pub fn get_data_objects(
&self
) -> impl Iterator<Item = (DataId, &DataDeclaration)>

Get an iterator of all data declarations

pub fn get_data_decl(&self, data_id: DataId) -> &DataDeclaration

Get the DataDeclaration for the data object named by name.

pub fn declare_function(
&mut self,
name: &str,
linkage: Linkage,
signature: &Signature
) -> ModuleResult<(FuncId, Linkage)>

Declare a function in this module.

pub fn declare_anonymous_function(
&mut self,
signature: &Signature
) -> ModuleResult<FuncId>

Declare an anonymous function in this module.

pub fn declare_data(
&mut self,
name: &str,
linkage: Linkage,
writable: bool,
tls: bool
) -> ModuleResult<(DataId, Linkage)>

Declare a data object in this module.

pub fn declare_anonymous_data(
&mut self,
writable: bool,
tls: bool
) -> ModuleResult<DataId>

Declare an anonymous data object in this module.