Crate cranelift_module

Top-level lib.rs for cranelift_module.

Structs

DataContext

This is to data objects what cranelift_codegen::Context is to functions.

DataDeclaration

Information about a data object which can be accessed.

DataDescription

A description of a data object.

DataId

A data object identifier for use in the Module interface.

FuncId

A function identifier for use in the Module interface.

FunctionDeclaration

Information about a function which can be called.

ModuleCompiledFunction

Information about the compiled function.

ModuleDeclarations

This provides a view to the state of a module which allows ir::ExternalNames to be translated into FunctionDeclarations and DataDeclarations.

ModuleReloc

A module relocation.

TrapSite

Record of the arguments cranelift passes to TrapSink::trap.

Enums

FuncOrDataId

A declared name may refer to either a function or data declaration

Init

This specifies how data is to be initialized.

Linkage

Linkage refers to where an entity is defined and who can see it.

ModuleError

Error messages for all Module methods

ModuleExtName

A translated ExternalName into something global we can handle.

Constants

VERSION

Version number of this crate.

Traits

Module

A Module is a utility for collecting functions and data objects, and linking them together.

Functions

default_libcall_names

Default names for [ir::LibCall]s. A function by this name is imported into the object as part of the translation of a [ir::ExternalName::LibCall] variant.

Type Definitions

ModuleResult

A convenient alias for a Result that uses ModuleError as the error type.