Enum cranelift_module::ModuleError

pub enum ModuleError {
    Undeclared(String),
    IncompatibleDeclaration(String),
    IncompatibleSignature(String, Signature, Signature),
    DuplicateDefinition(String),
    InvalidImportDefinition(String),
    Compilation(CodegenError),
    Allocation {
        message: &'static str,
        err: Error,
    },
    Backend(Error),
    Flag(SetError),
}

Error messages for all Module methods

Variants

Undeclared(String)

Indicates an identifier was used before it was declared

IncompatibleDeclaration(String)

Indicates an identifier was used as data/function first, but then used as the other

IncompatibleSignature(String, Signature, Signature)

Indicates a function identifier was declared with a different signature than declared previously

DuplicateDefinition(String)

Indicates an identifier was defined more than once

InvalidImportDefinition(String)

Indicates an identifier was defined, but was declared as an import

Compilation(CodegenError)

Wraps a cranelift-codegen error

Allocation

Fields

message: &'static str

Tell where the allocation came from

err: Error

Io error the allocation failed with

Memory allocation failure from a backend

Backend(Error)

Wraps a generic error from a backend

Flag(SetError)

Wraps an error from a flag definition.