Struct cranelift_module::DataContext

pub struct DataContext { /* private fields */ }

This is to data objects what cranelift_codegen::Context is to functions.

Implementations

impl DataContext

pub fn new() -> Self

Allocate a new context.

pub fn clear(&mut self)

Clear all data structures in this context.

pub fn define_zeroinit(&mut self, size: usize)

Define a zero-initialized object with the given size.

pub fn define(&mut self, contents: Box<[u8]>)

Define an object initialized with the given contents.

TODO: Can we avoid a Box here?

pub fn set_segment_section(&mut self, seg: &str, sec: &str)

Override the segment/section for data, only supported on Object backend

pub fn set_align(&mut self, align: u64)

Set the alignment for data. The alignment must be a power of two.

pub fn import_function(&mut self, name: ModuleExtName) -> FuncRef

Declare an external function import.

Users of the Module API generally should call Module::declare_func_in_data instead, as it takes care of generating the appropriate ExternalName.

pub fn import_global_value(&mut self, name: ModuleExtName) -> GlobalValue

Declares a global value import.

TODO: Rename to import_data?

Users of the Module API generally should call Module::declare_data_in_data instead, as it takes care of generating the appropriate ExternalName.

pub fn write_function_addr(&mut self, offset: CodeOffset, func: FuncRef)

Write the address of func into the data at offset offset.

pub fn write_data_addr(
&mut self,
offset: CodeOffset,
data: GlobalValue,
addend: Addend
)

Write the address of data into the data at offset offset.

pub fn description(&self) -> &DataDescription

Reference the initializer data.